Added MATLAB and Python files, see below
MATLAB and Python files for control have been added.  However, I am going
to make a major change to the whole architecture and replace everything
with a digital pattern generator.  In order to deal with the much larger
amount of data that might need to be sent, I will attempt to write data
via SPI instead of serial.